models/pyproject.toml
[project]
name = "open-fashion-clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "onnxscript>=0.3.0",
    "open-clip-torch>=2.32.0",
    "optimum>=1.25.3",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "ultralytics>=8.3.152",
]
